The problem is the initial zeros in lexicon Root:
LEXICON Root
0	nouns ;
... etc.

I removed the zeros, and inituppercase worked. Now, I do not know
why anyone would have such zeros, so in order not to break anything 
I instead inserted (0) in the inituppercase.regex:

... || .#. (0) _

and it works.

The same solution (include the flags in the regex) will work for
myv and izh initupper as well (or we might fix the lexc code).


git-svn-id: https://gtsvn.uit.no/langtech/trunk/langs/fin@64911 c7155fb1-f0a7-4240-a2fc-2600b6f42f90